rename NetteObjectToSmartTraitRector to generic ParentClassToTraitRector
