--- packages/Nette/tests/Rector/Utils/NetteObjectToSmartTraitRector/Correct/correct2.php.inc
@@ -1,7 +1,0 @@
-<?php declare (strict_types=1);
-
-class ClassWithExternalConstant
-{
-    use \Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\SomeTrait;
-    use AnotherTrait;
-}
