--- packages/Nette/tests/Rector/Utils/NetteObjectToSmartTraitRector/Correct/correct4.php.inc
@@ -1,9 +1,0 @@
-<?php declare (strict_types=1);
-
-use Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\ParentObject;
-
-class ClassWithExternalConstant
-{
-    use \Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\SomeTrait;
-    public $property;
-}
