--- packages/Nette/tests/Rector/Utils/NetteObjectToSmartTraitRector/Source/ParentObject.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source;
-
-class ParentObject
-{
-
-}
