--- packages/Nette/tests/Rector/Utils/NetteObjectToSmartTraitRector/Wrong/wrong2.php.inc
@@ -1,6 +1,0 @@
-<?php declare (strict_types=1);
-
-class ClassWithExternalConstant extends Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\ParentObject
-{
-    use AnotherTrait;
-}
