--- packages/Nette/tests/Rector/Utils/NetteObjectToSmartTraitRector/config.yml
@@ -1,5 +1,0 @@
-services:
-    Rector\Nette\Rector\Utils\NetteObjectToSmartTraitRector:
-        arguments:
-            $parentClass: 'Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\ParentObject'
-            $traitName: 'Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\SomeTrait'
