+++ tests/Rector/Class_/ParentClassToTraitRector/Correct/correct.php.inc
@@ -1,0 +1,8 @@
+<?php declare (strict_types=1);
+
+use Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject;
+
+class ClassWithExternalConstant
+{
+    use \Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\SomeTrait;
+}
