+++ tests/Rector/Class_/ParentClassToTraitRector/Correct/correct4.php.inc
@@ -1,0 +1,9 @@
+<?php declare (strict_types=1);
+
+use Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject;
+
+class ClassWithExternalConstant
+{
+    use \Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\SomeTrait;
+    public $property;
+}
