+++ tests/Rector/Class_/ParentClassToTraitRector/Source/ParentObject.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source;
+
+class ParentObject
+{
+
+}
