+++ tests/Rector/Class_/ParentClassToTraitRector/Source/SomeTrait.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source;
+
+trait SomeTrait
+{
+
+}
