+++ tests/Rector/Class_/ParentClassToTraitRector/Wrong/wrong.php.inc
@@ -1,0 +1,7 @@
+<?php declare (strict_types=1);
+
+use Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject;
+
+class ClassWithExternalConstant extends ParentObject
+{
+}
