+++ tests/Rector/Class_/ParentClassToTraitRector/Wrong/wrong2.php.inc
@@ -1,0 +1,8 @@
+<?php declare (strict_types=1);
+
+use Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject;
+
+class ClassWithExternalConstant extends ParentObject
+{
+    use AnotherTrait;
+}
