+++ tests/Rector/Class_/ParentClassToTraitRector/Wrong/wrong3.php.inc
@@ -1,0 +1,6 @@
+<?php declare (strict_types=1);
+
+class ClassWithExternalConstant extends Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject
+{
+    public $property;
+}
