+++ tests/Rector/Class_/ParentClassToTraitRector/Wrong/wrong4.php.inc
@@ -1,6 +1,6 @@
 <?php declare (strict_types=1);
 
-use Rector\Nette\Tests\Rector\Utils\NetteObjectToSmartTraitRector\Source\ParentObject;
+use Rector\Tests\Rector\Class_\ParentClassToTraitRector\Source\ParentObject;
 
 class ClassWithExternalConstant extends ParentObject
 {
